+++ app/scripts/lib/buy-eth-url.js
@@ -6,19 +6,8 @@ module.exports = {
 const ethNetProps = require('eth-net-props')
 
 const {
-  POA_CODE,
-  DAI_CODE,
-  POA_SOKOL_CODE,
-  MAINNET_CODE,
-  CLASSIC_CODE,
-  ROPSTEN_CODE,
-  RINKEBY_CODE,
-  KOVAN_CODE,
-  GOERLI_TESTNET_CODE,
   KARDIA_MAINNET_CODE,
-  KARDIA_TESTNET_CODE,
-  RSK_CODE,
-  RSK_TESTNET_CODE} = require('../controllers/network/enums')
+} = require('../controllers/network/enums')
 
 /**
  * Gives the caller a url at which the user can acquire coin, depending on the network they are in
@@ -35,26 +24,8 @@ const {
 function getBuyEthUrl ({ network, amount, address, ind }) {
   let url
   switch (Number(network)) {
-    case MAINNET_CODE:
-    case POA_CODE:
-    case DAI_CODE:
-    case CLASSIC_CODE:
-    case RSK_CODE:
-      url = getExchanges({network, amount, address})[ind].link
-      break
-    case ROPSTEN_CODE:
-    case RINKEBY_CODE:
-    case KOVAN_CODE:
-    case POA_SOKOL_CODE:
-    case RSK_TESTNET_CODE:
-    case GOERLI_TESTNET_CODE:
-      url = getFaucets(network)[ind]
-      break
     case KARDIA_MAINNET_CODE:
       url = ''
-      break
-    case KARDIA_TESTNET_CODE:
-      url = 'https://testnet.kardiachain.io/faucet'
       break
   }
   return url
@@ -82,60 +53,6 @@ function getFaucets (network) {
 function getExchanges ({network, amount, address}) {
   const networkID = Number(network)
   switch (networkID) {
-    case 1:
-      return [
-        {
-          link: `https://buy.coinbase.com/?code=9ec56d01-7e81-5017-930c-513daa27bb6a&amount=${amount}&address=${address}&crypto_currency=ETH`,
-        },
-      ]
-    case CLASSIC_CODE:
-      return [
-        {
-          name: 'Binance',
-          link: 'https://www.binance.com/en/trade/ETC_ETH',
-        },
-      ]
-    case POA_CODE:
-      return [
-        {
-          name: 'Binance',
-          link: 'https://www.binance.com/en/trade/POA_BTC',
-        },
-        {
-          name: 'BiBox',
-          link: 'https://www.bibox.com/exchange?coinPair=POA_ETH',
-        },
-        {
-          name: 'CEX Plus',
-          link: 'http://cex.plus/market/poa_eth',
-        },
-        {
-          name: 'HitBTC',
-          link: 'https://hitbtc.com/POA-to-ETH',
-        },
-      ]
-    case DAI_CODE:
-      return [
-        {
-          name: 'xDai TokenBridge',
-          link: 'https://dai-bridge.poa.network/',
-        },
-      ]
-    case RSK_CODE:
-      return [
-        {
-          name: 'Huobi',
-          link: 'https://www.huobi.com/',
-        },
-        {
-          name: 'Bitfinex',
-          link: 'https://www.bitfinex.com/',
-        },
-        {
-          name: 'Bitso',
-          link: 'https://bitso.com/',
-        },
-      ]
     case KARDIA_MAINNET_CODE:
       return [
         {